Started adding multiplayer in levels
- You can see other players that are in the same level as you
- Need to fix not being able to see other players when advancing to next level

+++ src/main/java/object/GameObject.java
@@ -17,7 +17,7 @@ public abstract class GameObject {
     public int width, height;
     public Color color;
     protected Color normalColor;
-    public byte direction;
+    public int direction;
     protected BufferedImage sprite;
 
     public GameObject(int x, int y, int width, int height, Color color) {
